reddit_scraping/requirements.txt
praw
streamlit==1.32.0
googletrans
python-dotenv
